iOS 16.0

+++ iOS/Developer/PrivateFrameworks/DVTFoundation.framework/NSError-DVTFoundationClassAdditions.h
@@ -11,10 +11,10 @@
 + (id)dvt_errorWithDomain:(id)arg1 errorCode:(long long)arg2 message:(id)arg3 recoverySuggestion:(id)arg4 recoveryOptions:(id)arg5 andRecoveryBlock:(CDUnknownBlockType)arg6;
 + (id)dvt_errorWithDomain:(id)arg1 errorCode:(long long)arg2 message:(id)arg3 recoverySuggestion:(id)arg4 underlyingError:(id)arg5;
 + (id)dvt_errorWithDomain:(id)arg1 errorCode:(long long)arg2 message:(id)arg3 recoverySuggestion:(id)arg4;
-+ (id)dvt_errorWithPOSIXErrorCode:(int)arg1;
-+ (id)dvt_errorWithPOSIXErrorCode:(int)arg1 messageFormat:(id)arg2;
-+ (id)dvt_errorWithDomain:(id)arg1 errorCode:(int)arg2 messageFormat:(id)arg3;
-+ (id)dvt_errorWithDomain:(id)arg1 errorCode:(int)arg2 messageFormat:(id)arg3 arguments:(char *)arg4;
++ (id)dvt_errorWithPOSIXErrorCode:(long long)arg1;
++ (id)dvt_errorWithPOSIXErrorCode:(long long)arg1 messageFormat:(id)arg2;
++ (id)dvt_errorWithDomain:(id)arg1 errorCode:(long long)arg2 messageFormat:(id)arg3;
++ (id)dvt_errorWithDomain:(id)arg1 errorCode:(long long)arg2 messageFormat:(id)arg3 arguments:(char *)arg4;
 + (id)dvt_errorWithSecurityErrorCode:(int)arg1;
 - (id)dvt_recursivelyRetrieveObjectForUserInfoKey:(id)arg1;
 - (id)dvt_propertyListDictionary;
