+++ iOS/PrivateFrameworks/Email.framework/EFContentProtectionObserver-Protocol.h
@@ -7,6 +7,6 @@
 #import <Email/NSObject-Protocol.h>
 
 @protocol EFContentProtectionObserver <NSObject>
-- (void)contentProtectionStateChanged:(int)arg1 previousState:(int)arg2;
+- (void)contentProtectionStateChanged:(long long)arg1 previousState:(long long)arg2;
 @end
