+++ iOS/PrivateFrameworks/Message.framework/EFContentProtectionObserver-Protocol.h
@@ -7,6 +7,6 @@
 #import <Message/NSObject-Protocol.h>
 
 @protocol EFContentProtectionObserver <NSObject>
-- (void)contentProtectionStateChanged:(int)arg1 previousState:(int)arg2;
+- (void)contentProtectionStateChanged:(long long)arg1 previousState:(long long)arg2;
 @end
